requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
playwright==1.53.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
